check if key exists inside of foreach loop for sample-template

+++ includes/templates/sample-templates/single.php
@@ -56,18 +56,19 @@ get_header(); ?>
 
 			<div class="entry-content">
 				<?php
-					foreach ($mbox_fields as $key => $value) {
+				foreach ($mbox_fields as $key => $value) {
 					$id = $value['id'];
+					if(array_key_exists($id,$meta)){
+						echo '<div class="team-item"><span class="team-item-label">';
+						echo $value['name'] . ': ';
+						echo '</span>';
 
-					echo '<div class="team-item"><span class="team-item-label">';
-					echo $value['name'] . ': ';
-					echo '</span>';
-
-					echo '<span class="team-item-value">';
-					echo $meta[$id][0];
-					echo '</span></div>';
-
-
+						echo '<span class="team-item-value">';
+						
+						echo $meta[$id][0];
+						
+						echo '</span></div>';
+					}
 				}
